refactor: split CLIAgent.ts (939 LOC) and APIAgent.ts (937 LOC) into focused sub-modules (#48)
CLIAgent split:
- src/agents/cli/types.ts - CLIAgentConfig, CLIProcessInfo, ExecutionContext, StreamData, DEFAULT_CLI_CONFIG
- src/agents/cli/CLICommandRunner.ts - process spawning, retry, kill, env management (217 LOC)
- src/agents/cli/CLIOutputParser.ts - output validation, wait-for-output, capture (168 LOC)
- src/agents/cli/index.ts - barrel exports
- src/agents/CLIAgent.ts - thin facade (185 LOC, down from 939)

APIAgent split:
- src/agents/api/types.ts - APIAgentConfig, HTTPMethod, AuthConfig, et al, DEFAULT_API_CONFIG
- src/agents/api/APIRequestExecutor.ts - Axios HTTP execution, retry, perf metrics (294 LOC)
- src/agents/api/APIResponseValidator.ts - status/header/body/schema validation (103 LOC)
- src/agents/api/APIAuthHandler.ts - bearer/apikey/basic auth (88 LOC)
- src/agents/api/index.ts - barrel exports
- src/agents/APIAgent.ts - thin facade (174 LOC, down from 937)

All public exports preserved for backward compatibility. tsc --noEmit passes cleanly.
