+++ src/agents/api/index.ts
@@ -1,0 +1,21 @@
+/**
+ * API sub-module barrel exports
+ */
+
+export type {
+  HTTPMethod,
+  AuthConfig,
+  RequestInterceptor,
+  ResponseInterceptor,
+  SchemaValidation,
+  PerformanceMeasurement,
+  RetryConfig,
+  APIAgentConfig,
+  APIRequest,
+  APIResponse,
+  RequestPerformance
+} from './types';
+export { DEFAULT_API_CONFIG } from './types';
+export { APIAuthHandler } from './APIAuthHandler';
+export { APIRequestExecutor } from './APIRequestExecutor';
+export { APIResponseValidator } from './APIResponseValidator';
