+++ src/agents/cli/index.ts
@@ -1,0 +1,13 @@
+/**
+ * CLI sub-module barrel exports
+ */
+
+export type {
+  CLIAgentConfig,
+  CLIProcessInfo,
+  ExecutionContext,
+  StreamData
+} from './types';
+export { DEFAULT_CLI_CONFIG } from './types';
+export { CLICommandRunner } from './CLICommandRunner';
+export { CLIOutputParser } from './CLIOutputParser';
